+++ src/TSTransformer/nodes/expressions/transformElementAccessExpression.ts
@@ -1,10 +1,13 @@
 import luau from "@roblox-ts/luau-ast";
 import { TransformState } from "TSTransformer";
+import { Prereqs } from "TSTransformer/classes/Prereqs";
 import { transformExpression } from "TSTransformer/nodes/expressions/transformExpression";
 import { transformOptionalChain } from "TSTransformer/nodes/transformOptionalChain";
 import { addIndexDiagnostics } from "TSTransformer/util/addIndexDiagnostics";
 import { addOneIfArrayType } from "TSTransformer/util/addOneIfArrayType";
 import { convertToIndexableExpression } from "TSTransformer/util/convertToIndexableExpression";
+import { tryMarkBuiltinMember } from "TSTransformer/util/evaluation/builtins";
+import { effectsCommute, getEffects, isLateRead, joinEffects, NO_EFFECTS } from "TSTransformer/util/evaluation/effects";
 import { getConstantValueLiteral } from "TSTransformer/util/getConstantValueLiteral";
 import { offset } from "TSTransformer/util/offset";
 import { skipUpwards } from "TSTransformer/util/traversal";
@@ -14,6 +17,7 @@ import ts from "typescript";
 
 export function transformElementAccessExpressionInner(
 	state: TransformState,
+	prereqs: Prereqs,
 	node: ts.ElementAccessExpression,
 	expression: luau.Expression,
 	argumentExpression: ts.Expression,
@@ -26,22 +30,28 @@ export function transformElementAccessExpressionInner(
 	const expType = state.typeChecker.getNonOptionalType(state.getType(node.expression));
 	addIndexDiagnostics(state, node, expType);
 
-	const [index, prereqs] = state.capture(() => transformExpression(state, argumentExpression));
+	const indexPrereqs = new Prereqs();
+	const index = transformExpression(state, indexPrereqs, argumentExpression);
 
-	if (!luau.list.isEmpty(prereqs)) {
+	if (
+		!effectsCommute(
+			getEffects(expression),
+			joinEffects(getEffects(indexPrereqs.statements), isLateRead(expression) ? getEffects(index) : NO_EFFECTS),
+		)
+	) {
 		// hack because wrapReturnIfLuaTuple will not wrap this, but now we need to!
 		if (isLuaTupleType(state)(expType)) {
 			expression = luau.array([expression]);
 		}
 
-		expression = state.pushToVar(expression, "exp");
-		state.prereqList(prereqs);
+		expression = prereqs.pushToVar(expression, "exp");
 	}
+	prereqs.pushList(indexPrereqs.statements);
 
 	// LuaTuple<T> checks
 	if (luau.isCall(expression) && isLuaTupleType(state)(expType)) {
 		// wrap in select() if it isn't the first value
-		if (!luau.isNumberLiteral(index) || Number(index.value) !== 0) {
+		if (!luau.isNumberLiteral(index) || Number(index.value.replace(/_/g, "")) !== 0) {
 			expression = luau.call(luau.globals.select, [offset(index, 1), expression]);
 		}
 		// parentheses to trim off the rest of the values
@@ -49,7 +59,7 @@ export function transformElementAccessExpressionInner(
 	}
 
 	if (ts.isDeleteExpression(skipUpwards(node).parent)) {
-		state.prereq(
+		prereqs.push(
 			luau.create(luau.SyntaxKind.Assignment, {
 				left: luau.create(luau.SyntaxKind.ComputedIndexExpression, {
 					expression: convertToIndexableExpression(expression),
@@ -62,17 +72,23 @@ export function transformElementAccessExpressionInner(
 		return luau.none();
 	}
 
-	return luau.create(luau.SyntaxKind.ComputedIndexExpression, {
+	const access = luau.create(luau.SyntaxKind.ComputedIndexExpression, {
 		expression: convertToIndexableExpression(expression),
 		index: addOneIfArrayType(state, expType, index),
 	});
+	tryMarkBuiltinMember(state, node, access);
+	return access;
 }
 
-export function transformElementAccessExpression(state: TransformState, node: ts.ElementAccessExpression) {
+export function transformElementAccessExpression(
+	state: TransformState,
+	prereqs: Prereqs,
+	node: ts.ElementAccessExpression,
+) {
 	const constantValue = getConstantValueLiteral(state, node);
 	if (constantValue) {
 		return constantValue;
 	}
 
-	return transformOptionalChain(state, node);
+	return transformOptionalChain(state, prereqs, node);
 }