+++ tests/src/diagnostics/noAny.7.ts
@@ -1,0 +1,5 @@
+function read(first: Array<any>, second: Array<any>) {
+	print(first[0]);
+	print(first[1]);
+	print(second[0]);
+}