+++ tests/src/helpers/util/prereqs/defaultValue.ts
@@ -1,0 +1,2 @@
+const values = [1, 2, 3];
+export default values.pop();